Add files via upload
This is the code of spam text analysis of a dataset, using this code we predict the low level text data of spam wepredict but the more confusing dataset is not predict by this because this code is train on Naive bayes Algorithm, with a very average accuracy on a small dataset. In this code, I create a Simple GUI by using tkinter module.